pyproject.toml
[tool.poetry]
name = "grpc-aiohttp"
version = "0.1.0"
description = ""
authors = ["Дмитрий Шойтов <shoytov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
mypy = "^0.942"
pep8 = "^1.7.1"
pylint = "^2.13.5"
black = "^22.3.0"
autoflake = "^1.4"
loguru = "^0.6.0"
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"
aiohttp = "^3.8.1"
aiohttp-swagger = "^1.0.16"
aiohttp-devtools = "^1.0.post0"
motor = "^3.0.0"
aiohttp-pydantic = "^1.12.2"
pydantic = {extras = ["email"], version = "^1.9.2"}
starlette = "^0.20.4"
PyJWT = "^2.4.0"
grpcio = "^1.48.0"
grpcio-tools = "^1.48.0"
googleapis-common-protos = "^1.56.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
